Fixed the BroadcastReceiver problem

+++ app/src/main/java/hua/dit/mobdev_project_2026/receiver/NotificationReceiver.java
@@ -13,10 +13,10 @@ public class NotificationReceiver extends BroadcastReceiver {
 
     @Override
     public void onReceive(Context context, Intent intent) {
-        long taskId = intent.getLongExtra("TASK_ID", -1);
+        int taskId = intent.getIntExtra("TASK_ID", -1);
 
         Log.d(TAG, "onReceive - Notification with id " + taskId + " is dismissed");
 
-        NotificationManagerCompat.from(context).cancel((int) taskId);
+        NotificationManagerCompat.from(context).cancel(taskId);
     }
 }
